requirements.txt
pygame-ce
asyncio
